Himanshu

+++ pecfestBackend/src/main/java/com/fest/pecfestBackend/service/EventService.java
@@ -1,6 +1,7 @@
 package com.fest.pecfestBackend.service;
 
 import com.fest.pecfestBackend.entity.Event;
+import com.fest.pecfestBackend.entity.User;
 import com.fest.pecfestBackend.enums.Club;
 import com.fest.pecfestBackend.enums.EventType;
 import com.fest.pecfestBackend.repository.EventRepo;
@@ -9,6 +10,7 @@ import com.fest.pecfestBackend.response.EventListByClubNameResponse;
 import com.fest.pecfestBackend.response.EventListResponse;
 import com.fest.pecfestBackend.response.TechnoCultEventResponse;
 import com.fest.pecfestBackend.response.WrapperResponse;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.http.HttpStatus;
@@ -21,16 +23,27 @@ import java.util.stream.Collectors;
 public class EventService {
     @Autowired
     private EventRepo eventRepo;
+    @Autowired
+    private SessionService sessionService;
 
     @Cacheable("eventsByClubName")
-    public WrapperResponse getEventsByOrganizingClubName(Club organizingClubName) {
-        return WrapperResponse.builder().data(eventRepo.findByOrganizingClub(organizingClubName)).build();
+    public WrapperResponse getEventsByOrganizingClubName(String organizingClubName) {
+        if(Club.fromString(organizingClubName)==null)
+            return WrapperResponse.builder().statusCode(HttpStatus.BAD_REQUEST.toString()).httpStatus(HttpStatus.BAD_REQUEST).statusMessage("No such club name exists!").build();
+        return WrapperResponse.builder().data(eventRepo.findAllByOrganizingClub(Club.fromString(organizingClubName))).build();
     }
 
-    public WrapperResponse addEvent(EventRequest addEventRequest) {
-        Event event = eventRepo.findByEventName(addEventRequest.getEventName());
-        if (Objects.isNull(event)) {
-            Event newEvent = getNewEvent(addEventRequest);
+    public WrapperResponse addEvent(EventRequest addEventRequest,String sessionId) {
+        User user=sessionService.verifySessionId(sessionId);
+        if(!Optional.ofNullable(user).isPresent()||Objects.isNull(user.getCoordinatingClubName())||user.getCoordinatingClubName().equals(Club.EMPTY))
+            return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized").build();
+
+        if(!user.getCoordinatingClubName().equals(Club.ALL)&&!user.getCoordinatingClubName().equals(Club.fromString(addEventRequest.getOrganizingClub())))
+            return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized to modify "+addEventRequest.getOrganizingClub()+" club").build();
+
+        List<Event> eventList = eventRepo.findAllByEventNameAndOrganizingClub(addEventRequest.getEventName(),Club.fromString(addEventRequest.getOrganizingClub()));
+        if (CollectionUtils.isEmpty(eventList)) {
+            Event newEvent = getNewEvent(addEventRequest,user.getName());
             eventRepo.save(newEvent);
             return WrapperResponse.builder().data(newEvent).statusMessage("Event added successfully").build();
         } else {
@@ -48,18 +61,25 @@ public class EventService {
                 workshop(workshopEventList).lecture(lectureEventList).build()).build();
     }
 
-    private Event getNewEvent(EventRequest addEventRequest) {
+    private Event getNewEvent(EventRequest addEventRequest,String name) {
         return Event.builder().eventCount(addEventRequest.getEventCount()).eventDescription(addEventRequest.getEventDescription()).eventEndDateAndTime(addEventRequest.getEventEndDateAndTime())
-                .organizingClub(addEventRequest.getOrganizingClub()).organizerContactNo(addEventRequest.getOrganizerContactNo())
+                .organizingClub(Club.fromString(addEventRequest.getOrganizingClub())).organizerContactNo(addEventRequest.getOrganizerContactNo())
                 .minNumberOfParticipants(addEventRequest.getMinNumberOfParticipants()).maxNumberOfParticipants(addEventRequest.getMaxNumberOfParticipants()).eventStartDateAndTime(addEventRequest.getEventStartDateAndTime())
                 .eventType(addEventRequest.getEventType()).prizeMoneyWorth(addEventRequest.getPrizeMoneyWorth())
-                .venue(addEventRequest.getVenue()).eventName(addEventRequest.getEventName()).rules(addEventRequest.getRules())
+                .venue(addEventRequest.getVenue()).eventName(addEventRequest.getEventName()).rules(addEventRequest.getRules()).eventBannerImageUrl(addEventRequest.getEventBannerImageUrl()).createdBy(name).updatedBy(name)
                 .build();
     }
 
-    public WrapperResponse deleteEvent(Long eventId) {
+    public WrapperResponse deleteEvent(Long eventId,String sessionId) {
         Optional<Event> event=eventRepo.findById(eventId);
         if(event.isPresent()) {
+            User user=sessionService.verifySessionId(sessionId);
+            if(!Optional.ofNullable(user).isPresent()||Objects.isNull(user.getCoordinatingClubName())||user.getCoordinatingClubName().equals(Club.EMPTY))
+                return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized").build();
+
+            if(!user.getCoordinatingClubName().equals(Club.ALL)&&!user.getCoordinatingClubName().equals(event.get().getOrganizingClub()))
+                return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized to modify "+event.get().getOrganizingClub()+" club").build();
+
             eventRepo.deleteById(eventId);
             return WrapperResponse.builder().data(event.get()).statusMessage("Event deleted successfully").build();
         }
@@ -68,16 +88,30 @@ public class EventService {
         }
     }
 
-    public WrapperResponse editEvent(Long eventId, EventRequest editEventRequest) {
+    public WrapperResponse editEvent(Long eventId, EventRequest editEventRequest,String sessionId) {
+        User user=sessionService.verifySessionId(sessionId);
+        if(!Optional.ofNullable(user).isPresent()||Objects.isNull(user.getCoordinatingClubName())||user.getCoordinatingClubName().equals(Club.EMPTY))
+            return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized").build();
+
+        if(!user.getCoordinatingClubName().equals(Club.ALL)&&!user.getCoordinatingClubName().equals(Club.fromString(editEventRequest.getOrganizingClub())))
+            return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized to modify "+editEventRequest.getOrganizingClub()+" club").build();
+
         Optional<Event> oldEventOptional=eventRepo.findById(eventId);
         if(oldEventOptional.isPresent()){
             Event oldEvent=oldEventOptional.get();
 
             oldEvent.setEventDescription(Optional.ofNullable(editEventRequest.getEventDescription()).orElse(oldEvent.getEventDescription()));
             oldEvent.setEventEndDateAndTime(Optional.ofNullable(editEventRequest.getEventEndDateAndTime()).orElse(oldEvent.getEventEndDateAndTime()));
-            oldEvent.setEventEndDateAndTime(Optional.ofNullable(editEventRequest.getEventStartDateAndTime()).orElse(oldEvent.getEventStartDateAndTime()));
+            oldEvent.setEventStartDateAndTime(Optional.ofNullable(editEventRequest.getEventStartDateAndTime()).orElse(oldEvent.getEventStartDateAndTime()));
             oldEvent.setVenue(Optional.ofNullable(editEventRequest.getVenue()).orElse(oldEvent.getVenue()));
             oldEvent.setOrganizerContactNo(Optional.ofNullable(editEventRequest.getOrganizerContactNo()).orElse(oldEvent.getOrganizerContactNo()));
+            oldEvent.setOrganizingClub(Club.fromString(editEventRequest.getOrganizingClub()));
+            oldEvent.setEventName(Optional.ofNullable(editEventRequest.getEventName()).orElse(oldEvent.getEventName()));
+            oldEvent.setEventCount(editEventRequest.getEventCount());
+            oldEvent.setMaxNumberOfParticipants(editEventRequest.getMaxNumberOfParticipants());
+            oldEvent.setMinNumberOfParticipants(editEventRequest.getMinNumberOfParticipants());
+            oldEvent.setEventType(editEventRequest.getEventType());
+            oldEvent.setRules(editEventRequest.getRules());
             eventRepo.save(oldEvent);
             return WrapperResponse.builder().data(oldEvent).statusMessage("Edited successfully").build();
         }
@@ -99,16 +133,28 @@ public class EventService {
     }
 
     private List<EventListByClubNameResponse> getListByClubName(List<Event> eventList) {
-        Map<Club,List<Event>> map=new HashMap<>();
+        Map<String,List<Event>> map=new HashMap<>();
         List<EventListByClubNameResponse> eventListByClubName=new ArrayList<>();
         for(Event event:eventList) {
-            if (!map.containsKey(event.getOrganizingClub()))
-                map.put(event.getOrganizingClub(), new ArrayList<>());
-            map.get(event.getOrganizingClub()).add(event);
+            if (!map.containsKey(event.getOrganizingClub().getClubName()))
+                map.put(event.getOrganizingClub().getClubName(), new ArrayList<>());
+            map.get(event.getOrganizingClub().getClubName()).add(event);
         }
         map.forEach((key,value)->{
             eventListByClubName.add(EventListByClubNameResponse.builder().clubName(key).eventList(value).build());
         });
         return eventListByClubName;
+    }
+
+    public WrapperResponse getEventsForClubAdmins(String sessionId) {
+        User user=sessionService.verifySessionId(sessionId);
+        if(!Optional.ofNullable(user).isPresent()||Objects.isNull(user.getCoordinatingClubName())||user.getCoordinatingClubName().equals(Club.EMPTY))
+            return WrapperResponse.builder().httpStatus(HttpStatus.FORBIDDEN).statusMessage("Not authorized").build();
+        else{
+            if(user.getCoordinatingClubName().equals(Club.ALL))// SuperAdmin
+                return WrapperResponse.builder().data(eventRepo.findAll()).build();
+            else
+                return WrapperResponse.builder().data(eventRepo.findAllByOrganizingClub(user.getCoordinatingClubName())).build();
+        }
     }
 }
